rename files to avoid weird diffs
